refactor: store 및 app provider 위치 변경
