+++ src/advanced/App.tsx
@@ -3,7 +3,7 @@ import GuideToggle from './components/guide/GuideToggle';
 import Header from './components/layout/Header';
 import Layout from './components/layout/Layout';
 import OrderSummary from './components/order/OrderSummary';
-import { AppProvider } from './lib/store';
+import { AppProvider } from './store';
 
 const App = () => {
   return (
